DAOS-6154 rdb, pool, container: Add layout versioning
To make future layout compatibility work slightly easier, this patch
adds versions to RDB, Pool Service, and Container Service layouts. For
each of these three layouts, one constant defines the default version
used normally for new instances, and another constant defines the
lowest compatible version. When an RDB, Pool Service, or Container
Service starts (or steps up, for the latter two), the layout version
of the persistent state is compared against the range defined by the
two constants, and -DER_DF_INCOMPA is returned if the version is not
in the range. (A notification should be sent to the MS upon
-DER_DF_INCOMPT. This will be implemented once the notification
mechanism is ready.)

Signed-off-by: Li Wei <wei.g.li@intel.com>