DAOS-6499 vos: save DTX modified obj in DRAM instead of PMEM
Originally, we saved the OIDs of the objects that are modified by
DTX in persistent memory. These OIDs will speed-up EC aggregation.
But EC aggregation does not depends on that, just optimization.
So saving them in DRAM is enough. Then it can reduce the space
overhead caused by the DTX entry (both actived and committed).

On the other hand, it also avoids some potential bugs caused by
reusing some committed DTX entry fields that was for saving the
objects' OIDs in PMEM.

Signed-off-by: Fan Yong <fan.yong@intel.com>