DAOS-6410 test: Placement Map Tests
Added several unit tests targeting the jump map placement algorithm.
Many of the tests are partially skipped because they uncover bugs
and are failing.

In order to take advantage of the the state change logic in
production code the ds_pool_map_tgts_update function was
abstracted out of the server library and created its own to be
linked to so that the placement tests could link to it
directly instead of needing to couple itself to the entire server.

Signed-off-by: Ryon Jensen <ryon.jensen@intel.com>